move pools to own file
